feat: 添加ora，整理cli.js的的初始化逻辑和目录结构

+++ src/init/husky.ts
@@ -1,24 +1,49 @@
 import path from "path";
 import fs from "fs-extra";
+import chalk from "chalk";
 import { execSync } from "child_process";
-function init() {
+
+import { CommanderOptions } from "../types";
+
+function init(options: CommanderOptions) {
   const huskyDirPath = path.join(process.cwd(), ".husky");
   const isInstalledHusky = fs.existsSync(huskyDirPath);
+  const isForce = options.force;
   if (!isInstalledHusky) {
     execSync("npx husky-init && npm install");
   }
-  updateHooks(huskyDirPath);
+  updateHooks(huskyDirPath, isForce);
 }
 
-function updateHooks(rootDir: string) {
+function updateHooks(rootDir: string, isForce: boolean) {
   const prepareCommitMsgPath = path.join(rootDir, "prepare-commit-msg");
   const commitMsgPath = path.join(rootDir, "commit-msg");
+
   if (fs.existsSync(prepareCommitMsgPath)) {
-    fs.removeSync(prepareCommitMsgPath);
+    if (isForce) {
+      fs.removeSync(prepareCommitMsgPath);
+    } else {
+      console.log(
+        chalk.yellowBright(`
+        The file already exists: .husky/prepare-commit-msg, husky is append, not override.
+        If you want to override hooks file, please use Use --force to override.
+        `),
+      );
+    }
   }
   if (fs.existsSync(commitMsgPath)) {
-    fs.removeSync(commitMsgPath);
+    if (isForce) {
+      fs.removeSync(commitMsgPath);
+    } else {
+      console.log(
+        chalk.yellowBright(`
+          The file already exists: .husky/commit-msg, husky is append , not override.
+          If you want to override hooks file, please use Use --force to override.
+          `),
+      );
+    }
   }
+
   // TODO 添加一些导语
   execSync("npx husky add .husky/prepare-commit-msg 'exec < /dev/tty && npx cz --hook || true'");
   execSync("npx husky add .husky/commit-msg 'npx commitlint --edit \"$1\"'");
